Fix native build: Gradle 8.7 + AGP 8.6.1, checksums, DNS responder, TCP proxy, adb-grantable private DNS
Verified: flutter analyze clean; debug + release APKs build (48MB).

+++ android/app/src/main/kotlin/com/kaya/booster/TcpProxy.kt
@@ -246,5 +246,5 @@ object TcpProxy {
         pool.shutdownNow()
     }
 
-    val activeConnections: Int get() = connections.get()
+    val activeConnections: Int get() = connections.get().toInt()
 }
